cloud-claude-service/requirements.txt
flask>=3.0.0
anthropic>=0.40.0
requests>=2.31.0
rich>=13.7.0
sseclient-py>=1.8.0
